🚜 Refactor the react-scripts utils
