+++ package.json
@@ -4,9 +4,9 @@
   "scripts": {
     "clean": "rm -rf node_modules; rm -rf packages/*/node_modules; rm yarn.lock; yarn cache clean",
     "lint": "node node_modules/eslint/bin/eslint.js packages",
-    "create-test-app": "rm -rf test-app; node packages/create-accurapp test-app --testing --no-install; cp -r .fixtures/test-app .; (cd test-app/; yarn)",
-    "create-test-app-build": "(yarn create-test-app; cd test-app/; yarn build)",
-    "create-test-app-start": "(yarn create-test-app; cd test-app/; yarn start)",
+    "create-test-app": "rm -rf test-app; node packages/create-accurapp test-app --testing --no-install && cp -r .fixtures/test-app . && (cd test-app/ && yarn)",
+    "create-test-app-build": "(yarn create-test-app && cd test-app/ && yarn build)",
+    "create-test-app-start": "(yarn create-test-app && cd test-app/ && yarn start)",
     "publish": "is-git-status-clean && lerna publish"
   },
   "workspaces": [
