+++ packages/accurapp-scripts/scripts/start.js
@@ -9,7 +9,14 @@ const detect = require('detect-port')
 const WebpackDevServer = require('webpack-dev-server')
 const openOrRefreshBrowser = require('react-dev-utils/openBrowser')
 const { prepareUrls } = require('react-dev-utils/WebpackDevServerUtils')
-const { log, createWebpackCompiler, readWebpackConfig, coloredBanner, extractBrowserslistString, extractLatestCommitHash, extractLatestCommitTimestamp, extractLatestTag } = require('./_utils')
+const { log, coloredBanner } = require('../utils/logging')
+const { createWebpackCompiler, readWebpackConfig } = require('../utils/webpack')
+const {
+  extractBrowserslistString,
+  extractLatestCommitHash,
+  extractLatestCommitTimestamp,
+  extractLatestTag,
+} = require('../utils/git')
 
 process.env.BROWSERSLIST = extractBrowserslistString()
 process.env.LATEST_COMMIT = extractLatestCommitHash()
@@ -50,12 +57,18 @@ function runDevServer(port) {
 
 console.log(coloredBanner('/||||/| accurapp'))
 
-detect(DEFAULT_PORT).then(port => {
-  if (port === DEFAULT_PORT) {
-    runDevServer(port)
-  } else {
-    log.ok(`Something is already running on port ${DEFAULT_PORT}, switching to ${chalk.blue(port)}...`)
-    runDevServer(port)
-  }
-  return port
-}).catch(err => { throw err })
+detect(DEFAULT_PORT)
+  .then(port => {
+    if (port === DEFAULT_PORT) {
+      runDevServer(port)
+    } else {
+      log.ok(
+        `Something is already running on port ${DEFAULT_PORT}, switching to ${chalk.blue(port)}...`
+      )
+      runDevServer(port)
+    }
+    return port
+  })
+  .catch(err => {
+    throw err
+  })
