+++ packages/create-accurapp/template/src/components/App.js
@@ -4,10 +4,7 @@ export default class App extends React.Component {
   render() {
     return (
       <div>
-        <h2>React App</h2>
-        <p>
-          To get started, edit <code>components/App.js</code> and save to reload.
-        </p>
+        <h2>Your React App goes here!</h2>
       </div>
     )
   }
