+++ packages/create-accurapp/template/src/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from 'components/App'
+import App from './components/App'
 import 'modern-normalize'
 import '@accurat/tachyons-lite'
 import 'tachyons-extra'
-import 'reset.css'
-import 'style.css'
+import './reset.css'
+import './style.css'
 
 function renderApp() {
   ReactDOM.render(<App />, document.getElementById('root'))
